sync

+++ libsrc/Wi/sparql.sql
@@ -7231,7 +7231,7 @@ create procedure DB.DBA.RDF_DELETE_TRIPLES_AGG (in graph_iid any, inout triples 
               declare search_fields_are_ok integer;
               search_fields_are_ok := __rdf_box_to_ro_id_search_fields (a_o, o_val, o_dt_and_lang_twobyte);
               -- dbg_obj_princ ('__rdf_box_to_ro_id_search_fields (', a_o, ') returned ', search_fields_are_ok, o_val, o_dt_and_lang_twobyte);
-	      if (__tag of rdf_box = __tag (a_o) and rdf_box_is_complete (a_o))
+              if (__tag of rdf_box = __tag (a_o) and rdf_box_ro_id (a_o)) -- was if (__tag of rdf_box = __tag (a_o) and rdf_box_is_complete (a_o))
                 delete from DB.DBA.RDF_QUAD where G = graph_iid and S = a_s and P = a_p and O = a_o;
 	      else if (search_fields_are_ok)
                 delete from DB.DBA.RDF_QUAD where G = graph_iid and S = a_s and P = a_p and O = (select rdf_box_from_ro_id(RO_ID) from DB.DBA.RDF_OBJ where RO_VAL = o_val and RO_DT_AND_LANG = o_dt_and_lang_twobyte);
@@ -7970,7 +7970,7 @@ grant select on DB.DBA.SPARQL_BINDINGS_VIEW to public
 ;
 
 --!AWK PUBLIC
-create procedure DB.DBA.SPARQL_BINDINGS_VIEW_C1_IMP (in dta any)
+create procedure DB.DBA.SPARQL_BINDINGS_VIEW_C_1_IMP (in dta any)
 {
   declare rcount, rctr integer;
   declare BND0 any;
@@ -7982,7 +7982,7 @@ create procedure DB.DBA.SPARQL_BINDINGS_VIEW_C1_IMP (in dta any)
 ;
 
 --!AWK PUBLIC
-create procedure DB.DBA.SPARQL_BINDINGS_VIEW_C2_IMP (in dta any)
+create procedure DB.DBA.SPARQL_BINDINGS_VIEW_C_2_IMP (in dta any)
 {
   declare rcount, rctr integer;
   declare BND0, BND1 any;
@@ -7994,7 +7994,7 @@ create procedure DB.DBA.SPARQL_BINDINGS_VIEW_C2_IMP (in dta any)
 ;
 
 --!AWK PUBLIC
-create procedure DB.DBA.SPARQL_BINDINGS_VIEW_C3_IMP (in dta any)
+create procedure DB.DBA.SPARQL_BINDINGS_VIEW_C_3_IMP (in dta any)
 {
   declare rcount, rctr integer;
   declare BND0, BND1, BND2 any;
@@ -8006,7 +8006,7 @@ create procedure DB.DBA.SPARQL_BINDINGS_VIEW_C3_IMP (in dta any)
 ;
 
 --!AWK PUBLIC
-create procedure DB.DBA.SPARQL_BINDINGS_VIEW_C4_IMP (in dta any)
+create procedure DB.DBA.SPARQL_BINDINGS_VIEW_C_4_IMP (in dta any)
 {
   declare rcount, rctr integer;
   declare BND0, BND1, BND2, BND3 any;
@@ -8017,28 +8017,28 @@ create procedure DB.DBA.SPARQL_BINDINGS_VIEW_C4_IMP (in dta any)
 }
 ;
 
-create procedure view DB.DBA.SPARQL_BINDINGS_VIEW_C1 as DB.DBA.SPARQL_BINDINGS_VIEW_C1_IMP (dta) (BND0 any)
+create procedure view DB.DBA.SPARQL_BINDINGS_VIEW_C_1 as DB.DBA.SPARQL_BINDINGS_VIEW_C_1_IMP (dta) (BND0 any)
 ;
 
-create procedure view DB.DBA.SPARQL_BINDINGS_VIEW_C2 as DB.DBA.SPARQL_BINDINGS_VIEW_C1_IMP (dta) (BND0 any, BND1 any)
+create procedure view DB.DBA.SPARQL_BINDINGS_VIEW_C_2 as DB.DBA.SPARQL_BINDINGS_VIEW_C_2_IMP (dta) (BND0 any, BND1 any)
 ;
 
-create procedure view DB.DBA.SPARQL_BINDINGS_VIEW_C3 as DB.DBA.SPARQL_BINDINGS_VIEW_C1_IMP (dta) (BND0 any, BND1 any, BND2 any)
+create procedure view DB.DBA.SPARQL_BINDINGS_VIEW_C_3 as DB.DBA.SPARQL_BINDINGS_VIEW_C_3_IMP (dta) (BND0 any, BND1 any, BND2 any)
 ;
 
-create procedure view DB.DBA.SPARQL_BINDINGS_VIEW_C4 as DB.DBA.SPARQL_BINDINGS_VIEW_C1_IMP (dta) (BND0 any, BND1 any, BND2 any, BND3 any)
+create procedure view DB.DBA.SPARQL_BINDINGS_VIEW_C_4 as DB.DBA.SPARQL_BINDINGS_VIEW_C_4_IMP (dta) (BND0 any, BND1 any, BND2 any, BND3 any)
 ;
 
-grant select on DB.DBA.SPARQL_BINDINGS_VIEW_C1 to public
+grant select on DB.DBA.SPARQL_BINDINGS_VIEW_C_1 to public
 ;
 
-grant select on DB.DBA.SPARQL_BINDINGS_VIEW_C2 to public
+grant select on DB.DBA.SPARQL_BINDINGS_VIEW_C_2 to public
 ;
 
-grant select on DB.DBA.SPARQL_BINDINGS_VIEW_C3 to public
+grant select on DB.DBA.SPARQL_BINDINGS_VIEW_C_3 to public
 ;
 
-grant select on DB.DBA.SPARQL_BINDINGS_VIEW_C4 to public
+grant select on DB.DBA.SPARQL_BINDINGS_VIEW_C_4 to public
 ;
